Add styling command line
Unfortunately I have to use this clunky hardcoded CSS for now as a
temporary workaround to style the command line to look like I want it to
look like. I need to investigate better non-deprecated method of doing
this, which would use the variables declared in `theme.cpp`. (And
eventually so that all of this can be overridden.)