Add crude hit-a-hint implementation
Hit-a-hint is a must have feature for a browser that is used mainly with
a keyboard. Implementing such feature with only C++ is impossible with
current WebKit API, due to DOM access being disabled, so I had to go
into my hacky mode and implement such thing with JavaScript included in
the C++ source code, which is then injected into the document.