Add concept of views
In order to implement split windows, we need to separate the web view
from the tab. This is where something called views. (In Vim, they would
be called windows but I decided to call them views so that it's less
confusing when working with GUI stuff.) A tab contains a view, which is
essentially the actual browser implementation. However, the concept is
designed so that there can be different kinds of views, ones for
displaying Web content and in the future ones which have been splitted
into multiple different views.